+++ src/plugins/blocksmith/components/blocksmith/index.jsx
@@ -2,6 +2,7 @@ import { EditorState } from 'draft-js';
 import React, { useState } from 'react';
 
 import Blocksmith from './Blocksmith';
+// eslint-disable-next-line import/no-named-as-default
 import ErrorBoundary from './error-boundary';
 
 export default ({ ...rest }) => {
